fix bug

+++ source/frame.cpp
@@ -3,4 +3,4 @@
 //
 #include "frame.h"
 #include <utility>
-Frame::Frame(int pts, int origin_pts, cv::Mat image) : pts_(pts), origin_pts_(origin_pts), image_(std::move(image)) {}
+Frame::Frame(uint64_t pts, uint64_t origin_pts, cv::Mat image) : pts_(pts), origin_pts_(origin_pts), image_(std::move(image)) {}
